// src/app/features/promotion-form/enum/tabs.enum.ts
export enum TabsEnum {
  DEFINITION = 'DEFINITION',
  CHOOSE_PRODUCT = 'CHOOSE_PRODUCT',
  EXCLUDE_PRODUCT = 'EXCLUDE_PRODUCT',
  BONUS_PRODUCTS = 'BONUS_PRODUCTS',
  PRODUCT_LIMITS = 'PRODUCT_LIMITS',
  CHOOSE_CLIENTS = 'CHOOSE_CLIENTS',
  EXCLUDE_CLIENTS = 'EXCLUDE_CLIENTS',
  CLIENTS_LIMIT = 'CLIENTS_LIMIT',
  SUMMARY = 'SUMMARY',
}